+++ packages/brick_offline_first/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Carry `providerArgs` from query when using `getBatched` (#200)
+
 ## 1.0.0
 
 * Null safety